Add endpoint that allows restarting pods by used image
