+++ routes/services.js
@@ -1,84 +1,8 @@
 const express = require('express')
 const router = express.Router()
-const spawnSync = require('child_process').spawnSync
 const debug = require('debug')('k8s-autodeploy:services')
-const config = require('../config')
-
-class NotSupportedDockerTagError extends Error {
-  constructor (message) {
-    super(message)
-    this.name = 'NotSupportedDockerTagError'
-  }
-}
-
-const executeCommand = (command, args, options = undefined) => {
-  let result = spawnSync(command, args, options)
-
-  if (result.status === 0) {
-    return Promise.resolve(result.stdout.toString())
-  } else {
-    return Promise.reject(result.stderr.toString())
-  }
-}
-
-const commonErrorHandler = (error, response, next) => {
-  if (error instanceof NotSupportedDockerTagError) {
-    debug('[WARNING] Reason => %s', error)
-    response.status(412).json({})
-  } else if (error.includes('NotFound')) {
-    // Replace `Error` word because it is a warning message
-    let warningMessage = error.replace('Error', '')
-
-    debug('[WARNING] %s', warningMessage)
-    response.status(412).json({})
-  } else {
-    // Default error handler
-    next(error)
-  }
-}
-
-const deployment = {
-  applyConfig (deploymentName, config) {
-    return executeCommand('kubectl', ['apply', '-f', '-'], {input: config})
-  },
-  delete (deploymentName) {
-    return executeCommand('kubectl', ['delete', 'deployment', deploymentName])
-  },
-  getConfig (deploymentName) {
-    return executeCommand('kubectl', ['get', 'deployment', deploymentName, '-o', 'yaml'])
-  },
-  getNameFromDockerTag (serviceName, dockerTag) {
-    switch (dockerTag) {
-      case 'develop':
-        return `${serviceName}`
-      case 'nightly':
-        return `dev-${serviceName}`
-      case 'release':
-        return `release-${serviceName}`
-      case 'staging':
-        return `staging-${serviceName}`
-      default:
-        // If config.ADDITIONAL_DOCKER_TAGS contains the docker tag
-        if(config.ADDITIONAL_DOCKER_TAGS.includes(dockerTag)) {
-          return `${dockerTag}-${serviceName}`
-        } else {
-          return null
-        }
-    }
-  },
-  restart (deploymentName) {
-    return this.getConfig(deploymentName).then(config => {
-      return this.delete(deploymentName).then(() => {
-        return this.applyConfig(deploymentName, config)
-      })
-    })
-  },
-  rollout (deploymentName) {
-    return this.getConfig(deploymentName).then(config => {
-      return executeCommand('kubectl', ['rollout', 'restart', 'deployment/' + deploymentName])
-    })
-  }
-}
+const { deployment } = require('../util/commands')
+const {NotSupportedDockerTagError, commonErrorHandler} = require('../util/errors');
 
 // Restart deployment
 router.post('/:services/restart', function (request, response, next) {
